# Reuse Deserializer::buf on every read_str and read_bin_data call