And a very synthetic benchmark:

```rust
#![feature(test)]

extern crate test;

extern crate serde;
extern crate rmp_serde;

use serde::{Serialize, Deserialize};

use test::{Bencher};

#[bench]
fn bench_strings_1000(bencher: &mut Bencher) {
    bench_strings(bencher, 1000)
}

#[bench]
fn bench_strings_5000(bencher: &mut Bencher) {
    bench_strings(bencher, 5000)
}

#[bench]
fn bench_strings_10000(bencher: &mut Bencher) {
    bench_strings(bencher, 10000)
}

fn bench_strings(bencher: &mut Bencher, size: usize) {
    let mut vec = Vec::new();
    for _ in 0..size {
        vec.push("abcdefghijklmnopqrstuvwxyz".to_string());
    }

    let mut buf = Vec::new();
    {
        let mut serializer = rmp_serde::Serializer::new(&mut buf);
        vec.serialize(&mut serializer).unwrap();
    }
    bencher.iter(|| {
        <Vec<String>>::deserialize(&mut rmp_serde::Deserializer::new(&buf[..])).unwrap();
    })
}
```

Before:

```
test bench_strings_1000  ... bench:     108,627 ns/iter (+/- 19,626)
test bench_strings_5000  ... bench:     544,226 ns/iter (+/- 32,176)
test bench_strings_10000 ... bench:   1,153,751 ns/iter (+/- 237,755)
```

After:

```
test bench_strings_1000  ... bench:      76,296 ns/iter (+/- 17,450)
test bench_strings_5000  ... bench:     387,087 ns/iter (+/- 45,780)
test bench_strings_10000 ... bench:     776,966 ns/iter (+/- 59,914)
```